apps/pyfunctions/requirements.txt
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.22.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytz==2024.2
requests==2.32.3
rsa==4.9
tbapy==1.3.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
